requirements.txt
fastapi
python-telegram-bot==13.15
uvicorn
scikit-learn
pandas
